[Bronze V] Title: Multiply, Time: 0 ms, Memory: 2020 KB, Score: 40 point -BaekjoonHub

+++ 백준/Bronze/22193. Multiply/Multiply.cc
@@ -2,8 +2,8 @@
 using namespace std;
 
 int main() {
-	int n, m;
-	int a, b;
+	long long int n, m;
+	long long int a, b;
 	cin >> n >> m >> a >> b;
 	cout << a * b;
 }
